Extract eco logc
